+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 ### New in 0.2 (not released yet)
 
-* _Nothing yet_
+* Breaking: Renamed `ConnectionString` to `MsSqlConnectionString`
 
 ### New in 0.1.8 (released 2015-09-16)